procedure to purge codegen caches; more optimizations as seen in profiling
- new PURGE_CODEGEN_CACHES procedure to clear codegeneration caches
- removed trove extensions THashMapWithCreate, TObjectLongHashMapWithIndex and replace
  with koloboke maps; also some usages of TObjectIntHashMap in TX code path replaced
- optimized CustomEntryConcurrentHashMap iteration by using a local expandible array
  for copying current chain instead of an ArrayList
- made ObjectObjectHashMap serializable by implementing Externalizable

+++ gemfire-core/src/main/java/com/gemstone/gemfire/internal/cache/ImageState.java
@@ -165,7 +165,7 @@ public interface ImageState /* extends Lock */ {
    * transaction has not been recorded as finished, and a negative of the order
    * if it was recorded as having been rolled back.
    */
-  public int getFinishedTXOrder(TXId txId);
+  public long getFinishedTXOrder(TXId txId);
 
   /**
    * Adjust the orders returned by getPendingTXOrder using the ordering from
